Pack more profiles in a row group.
The longitudinal dataset has only a few hundred profiles per row group,
which is detrimental for performance. This patch increases the number of
profiles per row group while simultaneously reducing the memory
footprint of the job, without which it wouldn't be possible to run on a typical c3.4xlarge cluster.