Add necessary fields for beta46 experiment 2
Starting from beta46 experiment 2, we're using the system addon mechanism to divide the beta and release populations into e10s test and non-e10s control groups. This allows us to perform more robust experimentsthan we could with, say, the Telemetry Experiments system.

That means the format of the experiments is now stored in environment/settings/e10sCohort. Links are available in the code with more context and details. The new e10s experiment format means that we need a new version and schema for the E10SExperiment derived streams.

Also, we removed old experiments, because we already have the data for those, and don't intend to re-run the jobs.