Add several more fields to main_summary v3.
This addresses bug 1270482, 1265874 and 1270505, issue #58 on
github, and adds the mean bookmarks and places counts.

We also change the LOOP_ACTIVITY_COUNTER data from a set of
individual fields to a single field containing a struct for all
the enum values (so that we can use the same approach for keyed
histograms too).