Implement an add-on recommendation engine.
This patch uses matrix factorization to recommend add-ons. The training
phase outputs a set of artifacts in JSON files which can be used from other
languages to compute a list of suggested add-ons given a set of installed
add-ons. The calculation requires a couple of matrix multiplications.
See `AddonRecommender.recommend` for a Scala implementation of how to do so.