Fetch addon info from addons.mozilla.org
Since clients may be sending an old name for an
addon, we need to make sure to get the current
name in the default locale for each addon. For
this reason, we fetch the addon info from
addons.mozilla.org to provide correct addon names
and also filter out unreviewed or private addons.