Store the AddonRecommender model to S3
AddonRecommender saves the generated data files
and model on the local machine. All the files
are uploaded to the provided S3 private bucket
(to keep an history of the generated data). The
data files are also mirrored in the public bucket,
to make them accessible over HTTP.
This also fixes the breakage due to the introduction
of the is_system property for the active_addons.