Add a basic CrossSectionalView with a single metric
This commit supports Bug 1286573 - Create cross-sectional parquet
dataset

This view only calculates a simple modal country and discards the data
after counting the number of rows. From this platform, we can save the
data to S3 and begin adding more metrics to meet the project
requirements.

There was a significant amount of troubleshooting required to get to
this stage. Accordingly, I'd like to keep this commit as the simplest
example of working with a Dataset instead of a Dataframe.