Replace case class to bypass the 22 field limit
Bug 1298123: Refactor dataset classes so they do not use case classes

Spark datasets can only be generated from classes which implement the
Product interface. We would usually use case classes for this purpose,
but case classes are limited to 22 fields. This change implements a
class to bypass this restriction.

This limit is removed in scala 2.11. Accordingly, this change should
only be a stop gap until we can upgrade our spark clusters.