Bug 1298763 - Create a derived parquet dataset for the sync ping.
This is a first cut that seems to "work", for some definitions of "work".

Alot of code was cargo-culted from MainSummaryView - I'm not sure how much
is relevant here. Some of the new parsing code I added is almost certainly
not idiomatic scala.

There are some tests. Some .json test files are parsed and their contents
verified. One of these payloads is also round-tripped through a local parquet
file and the content verified.

I also had a test that added a .heka file I grabbed of s3 and verified I could
parse it, but I felt that had little value as I couldn't actually verify the
contents. The commit where I removed it is at:
> https://github.com/mhammond/telemetry-batch-view/commit/ad40720d09d1717660bfc0ea140e67e80998f1b1
and I can reinstate that if you like.

Note that the code is handling the older format (where we sent one sync per
ping) and the current format (where we batch a number of syncs into one ping).
@Dexterp37 questioned whether the parquet structure should match the actual
ping formats, but I think this format is preferred - there's one row per
sync-event rather than one row per ping.