Get scalar data from clients with old pings
The scalars section was added by bug 1282091, which
landed on July 2016. Pings before that date don't
have the scalars section. If we have a ping with
the missing section among the pings for a client,
the longitudinal view drops all the data from that
client, if if newer pings have the scalar section.
This patch prevents that from happening.