Add a dimension to track profile age in CrashAggregateView.
We measure age in weeks with a min of -1 and a max of 53. Allowing unbounded
values is not necessary for the use case, and allows mischeivous clients to
grow our dataset unreasonably.

We suspect that crashiness in the first little while of a user's experience
with Firefox has a disproportionate effect on engagement and retention.

To test this hypothesis, we need to see how crashy things are based on the ages
of the profiles experiencing the crashes.