Use plain fields for browser engagement measures.
It appears that Presto doesn't like it when struct fields have their
inner structure changed (for example, by adding a new field). This
means we would have to back-fill data in order to add new nested
fields.

Instead, we add simple fields at the top level which do not cause
incompatibilities when new ones are added.