Bug 1369574 - Validate that searchAggregates are correct
This validates that search aggregates are working properly, with the
associated unit tests to verify that it is working. I thought that the
report was failing on calculating the search aggregates; it could have
and I've fixed it since, but the true problem lies somewhere with
createCrashReport.

This also changes Int -> Long, since that's the proper type as specified
by the main summary schema.