Add content histograms to longitudinal view
This change also makes it easy to add new histogram endpoints in the
future. Just add a path and suffix to histogramLocations or
keyedHistogramLocations. This whitelist updates both the parsing code
and the dataset schema.

I prototyped a version of this code which would remove the whitelist by
automatically scanning for new keys under "payload.processes". However,
we have to add these new keys/suffixes to the dataset schema.