Bug 1369574 - Add spark singleton and rewrite normalize function
The source of the bug ended up being in the way the spark session was
initialized. The spark session doesn't like being initialized in the
class scope and will causes all sorts of locking issues. Instead, a
singleton is created to prevent this issue from manifesting.

In addition, the normalize function I had previously defined did not
have good behavior on some input (its hard to tell when its running). I
replaced it with an implementation that is much easier to understand.

This also removes logging and uses println instead. Logging doesn't show
up where it's needed the most (the master node's cli).