Add ability to create pre-release Longitudinal
This covers quite a few changes:
1. Local files for scalars and histograms, and
tests for histograms.

2. Ability to specify which process will be used in the dataset.
That is, you can specify just content histograms.

3. Ability to specify which probe types - opt-in or opt-out.
It doesn't allow both right now.

4. Ability to choose different sources, samples, and channels
for data.

The reason we have these changes, rather than just including both
opt-in and opt-out together, is Spark broke on such a wide dataset.
Bringing it down to just one process type wasn't enough either; it
needs to be just opt-in for a process type to be feasible.

See bug 13490813 for discussion on this.