Bug 1314997 - Remove duplicate addon class definitions
Up to now, we just defined an addon class on-the-fly where we need it.
There are, in fact, two main use cases: one where we are reading addon-data
from a telemetry ping, another when we are serializing/deserializing data
from parquet format.

Let's create two versions of an Addon class, and use each where appropriate
(we can use the telemetry ping / camel case version for other use cases, like
addon analytics). The name of the parameters will be different in these cases,
to ease data manipulation (the json ping is camel case, parquet is snake case).