Bug 1329541: longitudinal datasource session_start_date is null for all records
This change adds session_start_date to the longitudinal dataset.

Note that this change removes the calls to `return` and consequentially
changes the behavior of the code in the case of missing
`subsession_start_date` values. However, these values are not missing in
practice, and this appears to be an undefined edge case currently.