Point users to Firefox Data Docs for end user docs
This change removes the high level user documentation from this
repository and directs readers to review the documentation here [1].

We leave the example commands for generating data and dataset schemata
in this repository. The example commands are too low level for the
unified documentation. The schemas can change frequently and should be
updated with the code.

I plan on dynamically including the dataset schemata in the reference
documentation. See https://bugzilla.mozilla.org/show_bug.cgi?id=1367135
for progress.

[1] https://mozilla.github.io/firefox-data-docs/concepts/choosing_a_dataset.html