Add Quantum Release Criteria view
This covers a few changes:
1. The Quantum RC view itself, which uses a new FilteredHLLMerge
UDAF.
2. The FilteredHLLMerge UDAF. This builds upon the previous
HllMerge by only adding a new client_id if some value is TRUE.
3. The Safe HyperLogLogMerge UDF. The previous version was
unsafe for NULL inputs, and unsafe when there were no rows
to aggregate (a NULL output).

The original plan was to use the GenericClientCount for this job,
but that soon became impossible due to the specific questions that
needed answered - e.g. sums, and Filtered HLLs.