Add a temporary list of disabled addon ids to main_summary
This field is required for the TAAR project and should not
be used for other purposes. See bug 1390814.
This makes use of the soon to be deprecated "addonDetails"
field that contains the list of the user's addons. It will
then populate the "disabled_addon_ids" column with the ids
of the addons that are in "addonDetails" but not in "activeAddons".