Fix the Addon HBase job from going OOM
Instead of applying a window function on the DataFrame
containing all the columns we require, we apply it to
a smaller DataFrame only containing the (client_id, subsession_start_date)
for each ping in the main_summary. We pick the most
recent entry and perform a join with the DataFrame
containing all the data we need to perform filtering.