Add ability to skip tests
This will support the quick-development use-case. Tests cannot
be skipped on tagged releases or master.

For example, if an engineer is working on a branch, and wants
to test out her changes using Airflow, they would need to build
the artifact and push it to s3 using a non-master branch. With
tests, that could take upwards of 1 hour -- with [skip-tests],
it should take under 5 minutes.

Note that for pr builds, this is always a merge commit; as such
they will always be tested (and why this commit is tested).