Update to spark 2.3.0
It's our understanding that APIs are backwards-compatible within
Spark 2.x, so taking the latest should be fine.

We also update the spark testing framework and make a variety
of necessary changes to deal with changes in libraries.

In particular, we add a security policy to handle a change in derby
and we standardize on DatasetSuiteBase for all tests that need Hive
support since we were incorrectly stopping the spark context, leading
to tests failing due to existing state from previous tests.