Use CircleCI 2.0 in place of TravisCI
This consolidates us down to one CI provider for t-b-v and takes
advantage of CircleCI's more powerful machinery for defining workflows
with parallel jobs and shared files between them.

We eliminate the Scalatest tags we previously used to manually assign
test cases to different build jobs. We now split up tests by source file
and use CircleCI's 'parallelize' config option to reliably assign them
to jobs.

We also eliminate the concept of git tags for controlling deploys,
as discussed during All Hands week. We always upload a successful
build as the current jar for the branch. This means now that whatever
hits `master` will be what's running in production.

The local workflow is simplified somewhat by no longer building a docker
image specific to this project, but rather relying on the mozilla/sbt
image.

This CircleCI configuration runs a bit faster and with better consistency
(9 to 10 minutes) compared to the current TravisCI configuration
(12 minutes to 20 minutes).