Bug 1473641 Set maxRecordsPerFile in ExperimentsSummaryView
Bug: https://bugzilla.mozilla.org/show_bug.cgi?id=1473641

The 'pref-hotfix-tls-13-avast-rollback' experiment has _much_ higher enrollment
than any other experiment in the past month.
It turned on on 6/28 with ~4 million appearances, which ballooned to 61M by 7/2.
This is at least an order of magnitude higher than any other experiment.

I'm pretty certain the problem we're hitting is that ExperimentSummaryView
is partitioning by experiment_id when writing with no cap on how
large a partition is, so it's trying to accumulate all 62M records
on a single node before writing out an object.

If we apply maxRecordsPerFile as we do on MainSummaryView,
then I think we'll be good.