Bug 1485682 - Scalarized Count and Flag histograms
Previously, we allowed Flag histograms (but not count), and they
were just stored as histograms (a map of buckets and their counts).

These two histogram types should get special treatment, since
at their heart they are scalars. This change allows people to
add count and flag histograms, but stores them as scalars.

Note that to do so requires a bit of processing of the histogram,
where previously we just stored whatever we found there. So for
example, a flag histogram with value 2 in the 0 bucket will be
a null, since that is not a valid representation. If we stored
that as a histogram, we would find just {0: 2}.

Finally we have one flag histogram which cannot be converted.
A new list of histogram-represented histograms is added,
and if a probe is present in that list it's exact `values`
will be included in main_summary.